Send delete request to backend

+++ src/components/user_detail/ConfirmationModal.js
@@ -1,12 +1,14 @@
 function ConfirmationModal(props) {
-    return (
-        <div className='modal'>
-            <p>Are you sure?</p>
-            <button className='btn' onClick={props.onCancel}>Cancel</button>
-            {/*<button className='btn' onClick={props.onDelete}>Confirm</button>*/}
-            <button className='btn' onClick={() => props.onDelete(props.id)}>Confirm</button>
-        </div>
-    );
+    if(props.isModalVisible) {
+        return (
+            <div className='modal'>
+                <p>Are you sure?</p>
+                <button className='btn' onClick={props.onCancel}>Cancel</button>
+                <button className='btn' onClick={() => props.onDelete(props.id)}>Confirm</button>
+            </div>
+        );
+    }
+    return <div/>
 }
 
 export {ConfirmationModal}
